First commit of many
What works: Modules, The new Store structure.
What doesn't work: CSS, socket, page serving (Now that modules are modular we have to figure out a dynamic id system).
