+++ app/css/style.css
@@ -1,467 +1,25 @@
-body {
-    font-family: "Lato", sans-serif;
-    -webkit-touch-callout: none;
-    -webkit-user-select: none;
-    -khtml-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-    overflow-x: hidden;
-    overflow-y: hidden;
-    /* cursor: default; */
-}
-
-.container {
-    padding-bottom: 20px;
-    display: flex;
-    font-size: 13px;
-    margin-left: -5px;
-}
-
-.container>div {
-    flex: 1;
-    /*grow*/
-    margin: 0 5px;
-}
-
-.container-title {
-    position: absolute;
-    left: 40px;
-    margin-top: 5px;
-}
-
-.mainMenu {
-    font-size: 10px;
-}
-
-.header {
-    height: 50px;
-    width: 250px;
-    position: fixed;
-    z-index: 2;
-    top: 0;
-    left: 60px;
-    overflow-x: hidden;
-    overflow-y: hidden;
-    -webkit-app-region: drag;
-    /* background-color: #fff; */
-}
-
-.sidenav {
-    height: 100%;
-    width: 60px;
-    position: fixed;
-    z-index: 1;
-    top: 0;
-    left: 0;
-    background-color: #1a1a1a;
-    overflow-x: hidden;
-    border-right: 1px solid #353535;
-    text-align: center;
-}
-
-.sidenav a:hover {
-    color: #f1f1f1;
-}
-
-.main {
-    font-size: 28px;
-    position: fixed;
-    top: 0;
-    right: 0;
-    margin-top: 50px;
-    color: #fff;
-    overflow-y: auto;
-    height: 87%;
-    width: 86%;
-    overflow-x: hidden;
-    margin-left: 20px;
-}
-
-.sub {
-    margin-left: 20px;
-}
-
-label {
-    padding: 0 0 5px 0;
-    font-size: 13px;
-    display: block;
-}
-
-h3 {
-    font-size: 20px;
-}
-
-h4 {
-    margin-top: 2px;
-    margin-bottom: 0;
-    font-size: 15px;
-    color: #FF5252;
-}
-
-.menuItem {
-    font-size: 13px;
-    margin: 0 0 0 0;
-}
-
-.menuItemDiv {
-    font-size: 13px;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
-
-hr {
-    display: block;
-    margin-top: 10px;
-    margin-bottom: 0.5em;
-    margin-left: auto;
-    margin-right: 7px;
-    border-style: inset;
-    border-width: 1px;
-    opacity: 0.3;
-}
-
-input[type=text] {
-    width: 130px;
-    height: 23px;
-    background-color: #1a1a1a;
-    color: #fff;
-    border: 1px solid #353535;
-    text-align: center;
-    font-family: "Lucida Console", Courier, monospace;
-    margin-bottom: 10px;
-}
-
-.container-title input[type=checkbox] {
-    position: absolute;
-    left: -20px;
-    top: -3px;
+html {
+    color: whitesmoke;
+    background-color: rgb(20, 20, 20);
 }
 
 .hide {
     display: none;
 }
 
-input[type=text]:focus {
-    outline: 0;
-}
-
-.button {
-    background-color: rgba(0, 0, 0, 0);
-    border: none;
-    color: white;
-    padding: 4px 0;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 12px;
-    cursor: default;
-    border-radius: 20px;
-    margin-top: 0;
-    border: 2px solid #a7a7a7 !important;
-    width: 80px;
-}
-
-.vol-slider {
-    background-color: rgba(0, 0, 0, 0);
-    border: none;
-    color: white;
-    padding: 4px 0;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 12px;
-    cursor: default;
-    border-radius: 20px;
-    margin-top: 0;
-    border: 2px solid #a7a7a7 !important;
-    width: 80px;
-}
-
-.buttonDiv {
-    margin-right: -30px;
-    margin-left: -30px;
-}
-
-.button:hover {
-    background-color: #FF5252;
-    color: #000;
-    border: 2px solid #FF5252 !important;
-}
-
-.button:active {
-    background-color: #fff;
-    color: #000;
-    border: 2px solid #fff !important;
-}
-
-.button:focus {
-    outline: 0;
-}
-
-/* 
-.generateButton {
-    background-color: #052AFF;
-    height: 40px;
-    width: 43%;
-    position: absolute;
-    z-index: 2;
-    bottom: 0;
-    left: 60px;
-    text-align: center;
-    vertical-align: middle;
-    line-height: 40px; 
-    text-decoration: none;
-    font-size: 12px;
-    color: #fff;
-} */
-
-table {
-    width: 70%;
-}
-
-table, th, td {
-    border: solid 1px #DDD;
-    border-collapse: collapse;
-    padding: 2px 3px;
-    text-align: center;
-}
-
-.tableFixHead {
-    overflow-y: auto;
-    height: 130px;
-}
-
-.tableFixHead thead th {
-    position: sticky;
-    top: 0;
-}
-
-/* Just common table stuff. Really. */
-
-table {
-    border-collapse: collapse;
-    width: 92%;
-}
-
-th, td {
-    padding: 8px 16px;
-    text-align: center;
-    border: solid 1px #DDD;
-}
-
-th {
-    background: rgb(56, 56, 56);
-    border: solid 1px #DDD;
-}
-
-.lowerButtonContainer {
-    display: flex;
-    width: 100%;
-    /* height: 40px; */
-    font-size: 12px;
-    text-align: center;
-    /* background-color: #052AFF; */
-    margin-top: 9px;
-    line-height: 40px;
-    /* position: absolute; */
-    margin-left: -10px;
-}
-
-.lowerButtonContainer>div {
-    flex: 1;
-    /*grow*/
-    border-right: 1px solid #c2c2c2;
-}
-
-.lowerButtonContainer>div:last-child {
-    border: none;
-}
-
-.lowerButton:hover {
-    background-color: #FF5252;
-    color: #000;
-}
-
-.lowerButton:active {
-    background-color: #fff;
-    color: #000;
-}
-
-@media screen and (max-height: 450px) {
-    .sidenav {
-        padding-top: 15px;
-    }
-    .sidenav a {
-        font-size: 18px;
-    }
-}
-
-#loading {
-    width: 30px;
-    position: fixed;
-    top: 10px;
-    right: 50px;
-}
-
-#closebtn {
-    font-size: 20px;
-    color: #fff;
-    position: fixed;
-    top: 16px;
-    right: 16px;
-    opacity: 0.3;
-}
-
-#closebtn:hover {
-    -webkit-filter: opacity(1) drop-shadow(0 0 0 #fff);
-    filter: opacity(1) drop-shadow(0 0 0 #fff);
-    opacity: 0.5;
-}
-
-#backbtn:hover {
-    -webkit-filter: opacity(1) drop-shadow(0 0 0 #fff);
-    filter: opacity(1) drop-shadow(0 0 0 #fff);
-    opacity: 0.5;
-}
-
-#backbtn {
-    font-size: 20px;
-    color: #fff;
-    position: fixed;
-    top: 16px;
-    right: 40px;
-    opacity: 0.3;
-}
-
-.title {
-    -webkit-filter: opacity(.5) drop-shadow(0 0 0 #fff);
-    filter: opacity(.5) drop-shadow(0 0 0 #fff);
-    position: fixed;
-    top: 18px;
-    left: 70px;
-}
-
-.sideNavItem {
-    cursor: default;
-    left: 17px;
-}
-
-#menuGoal {
-    bottom: 80px;
-}
-
-#menuSettings {
-    bottom: 30px;
-}
-
-#menuOverlay {
-    top: 80px;
-    left: 17px;
-}
-
-.menuIcon {
-    color: #c2c2c2;
-    font-size: 25px;
-    position: absolute;
-}
-
-.menuIcon:hover {
-    color: #FF5252;
-}
-
-a {
-    cursor: default;
-    outline: 0;
-}
-
-.active {
-    border: 3px solid #FF5252 !important;
-}
-
-.userIcon {
-    border-radius: 50%;
-    width: 45px;
-    height: 45px;
-    border: 3px solid #c2c2c2;
-    box-shadow: 0 0 1px 0px #c2c2c2 inset, 0 0 1px 0px #c2c2c2;
-    display: block;
-    margin: 0 auto;
-    object-fit: cover;
-}
-
-.userIcon:hover {
-    /* border: 3px solid #818181; */
-    border: 3px solid #FF5252 !important;
-    box-shadow: 0 0 1px 0px #FF5252 inset, 0 0 1px 0px #FF5252;
-}
-
-.userIcon.active {
-    border: 3px solid #FF5252 !important;
-    box-shadow: 0 0 1px 0px #FF5252 inset, 0 0 1px 0px #FF5252;
-}
-
-img {
-    -webkit-user-drag: none;
-    -khtml-user-drag: none;
-    -moz-user-drag: none;
-    -o-user-drag: none;
-    user-drag: none;
-    outline: 0;
-}
-
-#load {
-    margin-top: -25px;
-    margin-left: 20px;
-    width: 80px;
-    position: absolute;
-}
-
-#logo1 {
-    width: 150px;
-}
-
-#logo2 {
-    margin-top: -140px;
-    margin-left: -80px;
-    position: absolute;
-    width: 50px;
-}
-
-.centered {
-    position: fixed;
-    /* or absolute */
-    top: 150px;
-    left: 90px;
-}
-
-.slider {
-    width: 50px;
-    /* position: absolute;
-    margin-top: -4px;
-    margin-left: 5px;
-    margin-bottom: 0; */
-}
-
-/* width */
+/* Scrollbar */
 
 ::-webkit-scrollbar {
     width: 5px;
 }
 
-/* Track */
-
 ::-webkit-scrollbar-track {
     background: #aaaaaa;
 }
 
-/* Handle */
-
 ::-webkit-scrollbar-thumb {
     background: rgb(68, 68, 68);
 }
-
-/* Handle on hover */
 
 ::-webkit-scrollbar-thumb:hover {
     background: #555;
