+++ app/js_bk/renderer.js
@@ -1,0 +1,7 @@
+let closebtn = document.getElementById("closebtn");
+const path = require("path")
+
+closebtn.addEventListener("click", (e) => {
+    e.preventDefault();
+    window.closeCurrentWindow();
+});
