+++ run.bat
@@ -1,0 +1,5 @@
+@echo off
+
+cd app
+
+npm start
